doom: change 't_time.start' type
Variable 'start' is passed to as a pointer 'localtime()'. Function
'localtime()' accepts pointer to long int instead of long long int.
Thus, the type is changed to long int.

+++ game/inc/doom.h
@@ -332,7 +332,7 @@ typedef struct s_time
 	float			delta;
 	SDL_Color		color;
 	SDL_Surface		*surf;
-	long long int   start;
+	long int		start;
 	struct tm		date;
 	SDL_Color		clock_fg;
 	SDL_Color		clock_bg;
